+++ include/favsupport.h
@@ -13,6 +13,8 @@ submenu_item_t *readFavouritesFile(int *out_size);
 void addFavouriteItem(const submenu_item_t *item);
 void removeFavouriteByIdAndText(int id, const char *text);
 
+void loadFavourites(void);
+
 extern int gFAVStartMode;
 
 #endif
